Move stories helpers
