+++ packages/livechat/src/jpg.d.ts
@@ -1,0 +1,4 @@
+declare module '*.jpg' {
+	const path: string;
+	export = path;
+}
