+++ packages/livechat/src/mp3.d.ts
@@ -1,0 +1,4 @@
+declare module '*.mp3' {
+	const path: string;
+	export = path;
+}
