+++ packages/livechat/src/mp4.d.ts
@@ -1,0 +1,4 @@
+declare module '*.mp4' {
+	const path: string;
+	export = path;
+}
