+++ packages/livechat/src/png.d.ts
@@ -1,0 +1,4 @@
+declare module '*.png' {
+	const path: string;
+	export = path;
+}
