+++ packages/livechat/src/routes/Chat/stories.tsx
@@ -2,7 +2,7 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, Story } from '@storybook/preact';
 import type { ComponentProps } from 'preact';
 
-import { screenProps, avatarResolver, beepAudio, screenDecorator } from '../../helpers.stories';
+import { screenProps, avatarResolver, beepAudio, screenDecorator } from '../../../.storybook/helpers';
 import Chat from './component';
 
 const now = new Date(Date.parse('2021-01-01T00:00:00.000Z'));
