+++ packages/livechat/src/routes/Register/stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, Story } from '@storybook/preact';
 import type { ComponentProps } from 'preact';
 
 import Register from '.';
-import { screenDecorator, screenProps } from '../../helpers.stories';
+import { screenDecorator, screenProps } from '../../../.storybook/helpers';
 
 export default {
 	title: 'Routes/Register',
