+++ packages/livechat/src/routes/SwitchDepartment/stories.tsx
@@ -2,7 +2,7 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, Story } from '@storybook/preact';
 import type { ComponentProps } from 'preact';
 
-import { screenDecorator, screenProps } from '../../helpers.stories';
+import { screenDecorator, screenProps } from '../../../.storybook/helpers';
 import SwitchDepartment from './index';
 
 export default {
