+++ packages/livechat/tsconfig.json
@@ -6,7 +6,7 @@
 		"checkJs": false,
 		"jsxImportSource": "preact",
 	},
-	"include": ["./src", "./webpack.config.ts", "./svg-component-loader.ts"],
+	"include": ["./src", "./webpack.config.ts", "./svg-component-loader.ts", ".storybook/helpers.ts"],
 	"exclude": [
 		"./node_modules",
 		"./dist"
